pyproject.toml
[project]
name = "soict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arabic-reshaper==2.1.3",
    "beautifulsoup4>=4.13.5",
    "diffimg==0.2.3",
    "editdistance>=0.8.1",
    "jieba>=0.42.1",
    "logging>=0.4.9.6",
    "lxml>=6.0.1",
    "matplotlib>=3.10.6",
    "opencv-python>=4.2.0.32",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=10.0.0",
    "python-bidi==0.4.2",
    "python-levenshtein>=0.27.1",
    "requests>=2.20.0",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "torch>=2.8.0",
    "tqdm>=4.23.0",
    "transformers>=4.56.2",
    "wikipedia>=1.4.0",
]
